+++ entrypoint.sh
@@ -8,6 +8,9 @@ TZ=${TZ:-UTC}
 MEMORY_LIMIT=${MEMORY_LIMIT:-256M}
 UPLOAD_MAX_SIZE=${UPLOAD_MAX_SIZE:-16M}
 OPCACHE_MEM_SIZE=${OPCACHE_MEM_SIZE:-128}
+REAL_IP_FROM=${REAL_IP_FROM:-0.0.0.0/32}
+REAL_IP_HEADER=${REAL_IP_HEADER:-X-Forwarded-For}
+LOG_IP_VAR=${LOG_IP_VAR:-remote_addr}
 
 # Timezone
 echo "Setting timezone to ${TZ}..."
@@ -28,6 +31,9 @@ sed -e "s/@OPCACHE_MEM_SIZE@/$OPCACHE_MEM_SIZE/g" \
 # Nginx
 echo "Setting Nginx configuration..."
 sed -e "s/@UPLOAD_MAX_SIZE@/$UPLOAD_MAX_SIZE/g" \
+  -e "s#@REAL_IP_FROM@#$REAL_IP_FROM#g" \
+  -e "s#@REAL_IP_HEADER@#$REAL_IP_HEADER#g" \
+  -e "s#@LOG_IP_VAR@#$LOG_IP_VAR#g" \
   /tpls/etc/nginx/nginx.conf > /etc/nginx/nginx.conf
 
 # DokuWiki